backend/requirements.txt
Flask>=3.0,<4.0
flask-cors>=4.0,<5.0
pandas>=2.2,<3.0
numpy>=1.26,<3.0
scikit-learn>=1.5,<2.0
reportlab>=4.2,<5.0
python-dotenv>=1.0,<2.0
lime>=0.2.0.1,<0.3
